pyproject.toml
[project]
name = "dossier-review-ai-assistant"
version = "0.1.0"
description = "Local-first policy copilot for dossier review"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "pydantic>=2.8.0",
  "psutil>=7.0.0",
  "uvicorn>=0.30.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
